Fixed loss of data in form widget
