+++ lib/src/edit/widgets/widgets.dart
@@ -1,0 +1,1 @@
+export 'classe_info.dart';
