+++ lib/src/home/home.dart
@@ -1,4 +1,4 @@
-export 'components/components.dart';
+export 'widgets/widgets.dart';
 
 export 'home_controller.dart';
