+++ pubspec.yaml
@@ -8,33 +8,24 @@ version: 1.0.0+1
 environment:
   sdk: ">=2.7.0 <3.0.0"
 
-# dependency_overrides:
-#   analyzer: '0.39.17'
-#   build_resolvers: '1.3.10'
-
 dependencies:
   flutter:
     sdk: flutter
-
+  csv: ^4.1.0
   cupertino_icons: ^1.0.0
-
-  csv: ^4.0.3
-  provider: ^4.3.1
-  hive: ^1.4.3
-  hive_flutter: ^0.3.1
-  universal_html: ^1.2.3
-
   flutter_simple_treeview: ^2.0.2
+  hive: ^1.4.4+1
+  hive_flutter: ^0.3.1
+  provider: ^4.3.2+2
+  universal_html: ^1.2.3
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-
+  build_runner: ^1.10.4
   hive_generator: ^0.8.2
-  build_runner: ^1.10.1
 
 flutter:
   uses-material-design: true
   assets:
     - assets/gedalogo_270x270.png
-
